requirements.txt
ftfy
packaging
regex
tqdm
numpy==2.1.1
torch
torchvision
Pillow
scikit-learn
hnswlib
keyboard
deep-translator
PySimpleGUI==4.70.1
psutil